requirements.txt
gradio==4.44.0
langchain==0.3.9
langchain_community==0.3.9
langchain_core==0.3.21
langchain_ollama==0.2.1
langchain_openai==0.2.11
pymupdf==1.25.0
pymupdf4llm==0.0.17
python-dotenv==1.0.1
Requests==2.32.3